Update create-element to allow creation of Text nodes
carafe.dom/create-element was updated to allow for the creation of
Text nodes when it became clear that I could not create a button
with text that described what the button should do!

I updated the index.html file to allow me to experiment in the repl
to make sure that the function is working.

Also added :onclick event to know handlers.