Add carafe.file for browser based save functions
save-file is the base, and it uses the strategy of attaching an
'a' element to the document, then clicking it to trigger download.

save-svg makes it so that you don't have to remember the type to save
an svg file.

save-as-svg is the most convenient, you pass the element and the desired
filename you want and it will handle the download via the other functions.